Update JS-PRNG.js
Small changes in the seed value, e.g. +1 did not yield significantly different values, and patterns occurred. This was fixed on line 15. This is useful, for example, when requiring a certain coordinate to have an attached random value which would be achieved by setting the seed to x + y * someWidth, allowing the value for any coordinate to be obtained without looping through all the previous values to get to the correct state.